pyproject.toml
[project]
name = "ai-voc-analysis"
version = "0.1.0"
description = "AI VOC 선제 분석 — FastAPI + LangGraph + Splunk"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "gradio>=5.0",
  "uvicorn[standard]>=0.32",
  "langgraph>=0.2.60",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-anthropic>=0.3.0",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "python-dotenv>=1.0",
  "splunk-sdk>=2.0",
  "pyyaml>=6.0",
  "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = ["httpx>=0.27", "ruff>=0.8"]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
